+++ workdir/src/world/world.cpp
@@ -1,6 +1,4 @@
 #include<world/world.h>
-#include<cassert>
-#include<iostream>
 
 using namespace std;
 namespace world
@@ -50,7 +48,7 @@ namespace world
 
     world::~world()
     {
-        for (int i = 0; i < list_size; ++i)
+        for (size_t i = 0; i < list_size; ++i)
         {
             delete(object_list[i]);
         }
